Remove rack
Which we no longer need since we don't restart the coroutines for
transaction.

The transaction itself is (much) faster, which could be coroutine
and event overhead, and could also be confusing SQLite out of
applying write optimizations (such as batching subsequent writes
while the first one is blocked).

The asynchronous nature of the Lume will pull its weight over time,
but it's a good reminder.  The commit stage of the Lume is literally
a transaction, and the blocking code is easier to reason about and
faster as well.

The next step is to invert control over writes from the Skein
to the Lume.